load/unload model for all types of model
